Tests/LibWeb/Text/expected/hit_testing/adjacent-inline-links-fragment-hit-testing.txt

link-a
link-a
link-b
